Refactored things into multiple files for better readability
