+++ crates/bindings-typescript/src/server/index.ts
@@ -2,4 +2,4 @@ export * from './type_builders';
 export * from './type_util';
 export * from './schema';
 
-import './rt';
+import './runtime';
